% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_merge_billing.R
\name{clean_merge_billing}
\alias{clean_merge_billing}
\title{Clean & Merge Data Function}
\usage{
clean_merge_billing(dir = directory, first.day = date.first.day,
  date.current = date.current.report, final.report = F)
}
\description{
This function allows you import all of the associated files for one billing report, clean them, and merge them.
}
\examples{
clean_merge_billing()
}
\keyword{billing}
\keyword{cleaning}
\keyword{merging}
