% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_merge_cols.R
\name{get_merge_cols}
\alias{get_merge_cols}
\title{Get Merge Columns Function}
\usage{
get_merge_cols(data1, data2)
}
\description{
This is a generic function that identifies the common columns between 2 datasets for merging.
}
\examples{
get_merge_cols()
}
\keyword{cleaning}
\keyword{merging}
