% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_coach_report.R
\name{make_coach_report}
\alias{make_coach_report}
\title{Make Coach Report Function}
\usage{
make_coach_report(data, coach, dir = getwd(),
  date.current = date.current.report)
}
\description{
This function takes the output of the clean_merge_data function and produces coach reports as .csv files in the working directory.
}
\examples{
make_coach_report()
}
\keyword{coach}
\keyword{engagement}
\keyword{report}
