% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_coach_table.R
\name{make_coach_table}
\alias{make_coach_table}
\title{Make Coach Table Function}
\usage{
make_coach_table(dir = directory, data,
  current.report = date.current.report)
}
\description{
This function takes the output of the clean_merge_data function and computes coach capacity and related statistics.
}
\examples{
make_coach_t()
}
\keyword{coach}
\keyword{table}
