% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stats_table.R
\name{make_stats_table}
\alias{make_stats_table}
\title{Make Stats Table Function}
\usage{
make_stats_table(dir = directory, data,
  date.current = date.current.report, saas = F)
}
\description{
This function takes the output of the clean_merge_data function and turns it into the customer statistics table.
}
\examples{
make_stats_table()
}
\keyword{coach}
\keyword{customer}
\keyword{table}
