% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_coach_report.R
\name{make_coach_report}
\alias{make_coach_report}
\title{Make Coach Engagement Reports}
\usage{
make_coach_report(data, coach, dir = getwd(),
  date.current = date.current.report)
}
\description{
This function takes a cleaned billing file and creates the weekly engagement report for a specific coach.
}
\examples{
make_coach_report()
}
\keyword{coach}
\keyword{engagement}
\keyword{report}
